+++ app/components/projects/new_component.html.erb
@@ -39,7 +39,7 @@ See COPYRIGHT and LICENSE files for more details.
           render(
             Primer::Forms::FormList.new(
               Projects::Settings::NameForm.new(f),
-              Projects::Settings::EditableIdentifierForm.new(f),
+              Projects::Settings::IdentifierForm.new(f),
               Projects::Settings::DescriptionForm.new(f),
               Projects::Settings::RelationsForm.new(f, invisible: params[:parent_id].present?),
               Projects::Settings::TypeForm.new(f)
