--- app/forms/projects/settings/editable_identifier_form.rb
@@ -1,60 +1,0 @@
-# frozen_string_literal: true
-
-#-- copyright
-# OpenProject is an open source project management software.
-# Copyright (C) the OpenProject GmbH
-#
-# This program is free software; you can redistribute it and/or
-# modify it under the terms of the GNU General Public License version 3.
-#
-# OpenProject is a fork of ChiliProject, which is a fork of Redmine. The copyright follows:
-# Copyright (C) 2006-2013 Jean-Philippe Lang
-# Copyright (C) 2010-2013 the ChiliProject Team
-#
-# This program is free software; you can redistribute it and/or
-# modify it under the terms of the GNU General Public License
-# as published by the Free Software Foundation; either version 2
-# of the License, or (at your option) any later version.
-#
-# This program is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# GNU General Public License for more details.
-#
-# You should have received a copy of the GNU General Public License
-# along with this program; if not, write to the Free Software
-# Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
-#
-# See COPYRIGHT and LICENSE files for more details.
-#++
-module Projects
-  module Settings
-    class EditableIdentifierForm < ApplicationForm
-      form do |f|
-        if Setting::WorkPackageIdentifier.alphanumeric?
-          f.text_field(
-            name: :identifier,
-            label: attribute_name(:identifier),
-            caption: I18n.t("projects.settings.change_identifier_format_hint_semantic"),
-            required: true,
-            validation_message: validation_message_for(:identifier)
-          )
-        else
-          f.text_field(
-            name: :identifier,
-            label: attribute_name(:identifier),
-            caption: I18n.t("projects.settings.change_identifier_format_hint_legacy"),
-            required: true,
-            validation_message: validation_message_for(:identifier)
-          )
-        end
-      end
-
-      private
-
-      def validation_message_for(attribute)
-        model.errors.full_messages_for(attribute).to_sentence.presence
-      end
-    end
-  end
-end
